adk-agents-projects/rag-gemini-search-files/requirements.txt
google-genai==1.49.0
google-adk==1.18.0
rich==13.7.1
httpx==0.28.1
tenacity==9.0.0
PyYAML==6.0.2
